+++ components/src/core/index.ts
@@ -10,3 +10,5 @@ export * from './ScoreCard';
 export * from './Utility';
 export * from './UserMenu';
 export * from './DropdownToolbar';
+export * from './AppBar';
+export * from './ThreeLiner';