+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 5.3.0 (June 30, 2021)
+
+### Added
+
+-   Added `<AppBar>` component which will expand and collapse based on scroll position.
+-   Added `wrapInfo` property onto `<InfoListItem>`.
+
 ## 5.2.0 (May 28, 2021)
 
 ### Changed